+++ packages/client/src/StreamrClient.ts
@@ -27,6 +27,7 @@ import { Methods, Plugin } from './utils/Plugin'
 import { StreamDefinition } from './types'
 import { Subscription, SubscriptionOnMessage } from './subscribe/Subscription'
 import { StreamIDBuilder } from './StreamIDBuilder'
+import { EventEmitterInjectionToken, StreamrClientEventEmitter, StreamrClientEvents } from './events'
 
 let uid: string = process.pid != null
     // Use process id in node uid.
@@ -80,7 +81,8 @@ class StreamrClientBase implements Context {
         private dataunions: DataUnions,
         private streamRegistry: StreamRegistry,
         private storageNodeRegistry: StorageNodeRegistry,
-        private streamIdBuilder: StreamIDBuilder
+        private streamIdBuilder: StreamIDBuilder,
+        private eventEmitter: StreamrClientEventEmitter
     ) { // eslint-disable-line function-paren-newline
         this.id = context.id
         this.debug = context.debug
@@ -96,6 +98,7 @@ class StreamrClientBase implements Context {
         Plugin(this, this.dataunions)
         Plugin(this, this.streamRegistry)
         Plugin(this, this.storageNodeRegistry)
+        Plugin(this, this.eventEmitter)
 
         this.onDestroy = this.destroySignal.onDestroy.bind(this.destroySignal)
         this.isDestroyed = this.destroySignal.isDestroyed.bind(this.destroySignal)
@@ -170,6 +173,18 @@ class StreamrClientBase implements Context {
     disableDebugLogging() { // eslint-disable-line class-methods-use-this
         Debug.disable()
     }
+
+    on<T extends keyof StreamrClientEvents>(eventName: T, listener: StreamrClientEvents[T]) {
+        this.eventEmitter.on(eventName, listener as any)
+    }
+
+    once<T extends keyof StreamrClientEvents>(eventName: T, listener: StreamrClientEvents[T]) {
+        this.eventEmitter.once(eventName, listener as any)
+    }
+
+    off<T extends keyof StreamrClientEvents>(eventName: T, listener: StreamrClientEvents[T]) {
+        this.eventEmitter.off(eventName, listener as any)
+    }
 }
 
 /**
@@ -177,6 +192,8 @@ class StreamrClientBase implements Context {
  */
 export function initContainer(config: StrictStreamrClientConfig, parentContainer = rootContainer) {
     const c = parentContainer.createChildContainer()
+    const eventEmitter = new StreamrClientEventEmitter()
+    c.register(EventEmitterInjectionToken, { useValue: eventEmitter } as any)
     uid = uid || `${uuid().slice(-4)}${uuid().slice(0, 4)}`
     const id = counterId(`StreamrClient:${uid}${config.id ? `:${config.id}` : ''}`)
     const debug = Debug(id)
@@ -252,7 +269,8 @@ export class StreamrClient extends StreamrClientBase {
             c.resolve<DataUnions>(DataUnions),
             c.resolve<StreamRegistry>(StreamRegistry),
             c.resolve<StorageNodeRegistry>(StorageNodeRegistry),
-            c.resolve<StreamIDBuilder>(StreamIDBuilder)
+            c.resolve<StreamIDBuilder>(StreamIDBuilder),
+            c.resolve(EventEmitterInjectionToken)
         )
     }
 }