+++ packages/client/test/test-utils/fake/FakeStorageNodeRegistry.ts
@@ -4,13 +4,14 @@ import { StreamIDBuilder } from '../../../src/StreamIDBuilder'
 import { DOCKER_DEV_STORAGE_NODE } from '../../../src/ConfigTest'
 import { FakeStorageNode } from './FakeStorageNode'
 import { ActiveNodes } from './ActiveNodes'
-import { StorageNodeAssignmentEvent, StorageNodeRegistry } from '../../../src/StorageNodeRegistry'
+import { StorageNodeRegistry } from '../../../src/StorageNodeRegistry'
 import { Stream } from '../../../src/Stream'
 import { Multimap } from '../utils'
 import { StreamRegistry } from '../../../src/StreamRegistry'
 
 @scoped(Lifecycle.ContainerScoped)
 export class FakeStorageNodeRegistry implements Omit<StorageNodeRegistry,
+    'initStreamAssignmentEventListener' |
     'clientConfig' | 'chainProvider' | 'streamStorageRegistryContractReadonly' |
     'chainSigner' | 'nodeRegistryContract' | 'streamStorageRegistryContract'> {
 
@@ -108,16 +109,6 @@ export class FakeStorageNodeRegistry implements Omit<StorageNodeRegistry,
 
     // eslint-disable-next-line class-methods-use-this
     getStoredStreams(_nodeAddress: EthereumAddress): Promise<{ streams: Stream[]; blockNumber: number }> {
-        throw new Error('not implemented')
-    }
-
-    // eslint-disable-next-line class-methods-use-this
-    registerStorageEventListener(_listener: (event: StorageNodeAssignmentEvent) => any): Promise<void> {
-        throw new Error('not implemented')
-    }
-
-    // eslint-disable-next-line class-methods-use-this
-    unregisterStorageEventListeners(): Promise<void> {
         throw new Error('not implemented')
     }
 }